removed extra service directory after moving mockdata and api services to service directory under core
