+++ frontend/nafis-frontend-app/src/app/core/services/api.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import {ApiResponse, FiltreRecherche, Pagination} from "../interfaces/common";
-import {Rendezvous} from "../interfaces/rendezvous";
-import {Personnel} from "../interfaces/personnel";
-import {Chambre} from "../interfaces/chambre";
-import {ConstantesVitales} from "../interfaces/surveillance";
-import {Patient} from "../interfaces/patient";
+import {ApiResponse, FiltreRecherche, Pagination} from "../../interfaces/common";
+import {Rendezvous} from "../../interfaces/rendezvous";
+import {Personnel} from "../../interfaces/personnel";
+import {Chambre} from "../../interfaces/chambre";
+import {ConstantesVitales} from "../../interfaces/surveillance";
+import {Patient} from "../../interfaces/patient";
 
 
 @Injectable({
