+++ frontend/nafis-frontend-app/src/app/core/services/mock-data.service.spec.ts
